+++ apps/web/src/components/task-list-sidebar.tsx
@@ -10,6 +10,7 @@ import { useTasks, ThreadSummary } from "@/providers/Task";
 import { useQueryState, parseAsString } from "nuqs";
 import { useEffect, useState } from "react";
 import { ThreadItem } from "./thread-item";
+import { groupTasksIntoThreads, sortThreadsByDate } from "@/lib/thread-utils";
 
 const THREADS_PER_PAGE = 10; // More threads per page in sidebar
 
@@ -40,56 +41,11 @@ export default function TaskListSidebar({ onCollapse }: TaskListSidebarProps) {
     setTaskId(null); // Clear task selection to show thread in chat mode
   };
 
-  // Group tasks by thread
-  const threadSummaries: ThreadSummary[] = allTasks.reduce((acc, task) => {
-    const existingThread = acc.find((t) => t.threadId === task.threadId);
+  // Group tasks by thread using utility function
+  const threadSummaries = groupTasksIntoThreads(allTasks);
 
-    if (existingThread) {
-      existingThread.tasks.push(task);
-      existingThread.totalTasksCount += 1;
-      if (task.status === "done") {
-        existingThread.completedTasksCount += 1;
-      }
-    } else {
-      acc.push({
-        threadId: task.threadId,
-        threadTitle:
-          task.threadTitle || `Thread ${task.threadId.substring(0, 8)}`,
-        repository: task.repository || "Unknown Repository",
-        branch: task.branch || "main",
-        date: task.date,
-        createdAt: task.createdAt,
-        tasks: [task],
-        completedTasksCount: task.status === "done" ? 1 : 0,
-        totalTasksCount: 1,
-        status: task.status,
-      });
-    }
-
-    return acc;
-  }, [] as ThreadSummary[]);
-
-  // Determine overall thread status
-  threadSummaries.forEach((thread) => {
-    const hasRunning = thread.tasks.some((t) => t.status === "running");
-    const hasError = thread.tasks.some((t) => t.status === "error");
-    const allDone = thread.tasks.every((t) => t.status === "done");
-
-    if (hasError) {
-      thread.status = "error";
-    } else if (hasRunning) {
-      thread.status = "running";
-    } else if (allDone) {
-      thread.status = "done";
-    } else {
-      thread.status = "interrupted";
-    }
-  });
-
-  // Sort threads by creation date (newest first)
-  const sortedThreads = threadSummaries.sort(
-    (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime(),
-  );
+  // Sort threads by creation date (newest first) using utility function
+  const sortedThreads = sortThreadsByDate(threadSummaries);
 
   const totalThreads = sortedThreads.length;
   const totalPages = Math.ceil(totalThreads / THREADS_PER_PAGE);
